+++ CHANGELOG.rst
@@ -53,9 +53,13 @@ Version 1.5.1
 
     res = client.read_holding_registers(...)
     if not res.isError():
+
         # proceed
+ 
     else:
         # handle error or raise
+
+    """
 
 * Add examples for MEI read device information request